src/UserService/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
pydantic
PyJWT
passlib[bcrypt]
pydantic[email]
cryptography
kafka-python==2.0.2